split src/lib.rs into focused modules (closes #4)
Moves parser, serde, cmp, and convert code out of lib.rs (502 -> 65 lines).
Public API unchanged; all 70 tests + 5 doctests pass under both feature configs.
